# speed up encode