@michaelkirk 2024-05-01
blech, sorry, I missed https://github.com/georust/polyline/pull/40. Let me look at that first.