@mattiZed 2024-05-01
IMHO it should be easy to combine this change with #40 :)